Refactor cluster creation.
